feature(file names): It's now possible to change file names.
To fix the styles for the input field, I happened to stumble upon the
browser native styles, which told me exactly how to solve the problem.
I'll try to use this approach in the future as well.

Here's the URL to Firefox's default CSS:
view-source:resource://gre-resources/forms.css helped